Add checking for row and columns
Finished adding a function that checks if the user or if the cpu has
complete a row or column. Also added a screen that will show the user
whether they won, lost of if the match was a draw. I still need to add
diagonal checking, but the project is a lot more functional now.